refactor: move handleSnapshot and handleSnapshotCheck to solidity-test/task-action
